Change the project to make 10.6 & retina compatible icons.
The .icns files generated by iconutil from a .iconset
have 128x128 icons which are not usable in OS X 10.6.

They have some kind of type in the .icns file of "ic07"
when generated by iconutil, but that was a "virtual type"?
or something in the 10.6 SDK, and not really present on disk?

So a different format is used there, and SetIconFamilyData()
appears to be able to create that format.

So this just mods main.c to ignore all icon formats other than
"ic07" and write them out just the way they were in the input file.

For the "ic07" type, it will generate a 10.6 appropriate 128x128.

Ken told me how to do this & I cribbed the code from some he'd written.